+++ internal/app/usecase/auth/register.go
@@ -3,8 +3,8 @@ package auth
 import (
 	"fmt"
 
-	"github.com/Marvit-Solutions/csw-golang/internal/domain/model/request"
-	"github.com/Marvit-Solutions/csw-golang/internal/domain/model/response"
+	"github.com/Marvit-Solutions/csw-golang/internal/domain/localmodel/request"
+	"github.com/Marvit-Solutions/csw-golang/internal/domain/localmodel/response"
 	"github.com/Marvit-Solutions/csw-golang/library/config"
 	"github.com/Marvit-Solutions/csw-golang/library/middleware/auth"
 	"github.com/Marvit-Solutions/csw-golang/library/struct/model"
