+++ internal/app/usecase/module/material_all.go
@@ -3,8 +3,8 @@ package module
 import (
 	"fmt"
 
-	"github.com/Marvit-Solutions/csw-golang/internal/domain/model/request"
-	"github.com/Marvit-Solutions/csw-golang/internal/domain/model/response"
+	"github.com/Marvit-Solutions/csw-golang/internal/domain/localmodel/request"
+	"github.com/Marvit-Solutions/csw-golang/internal/domain/localmodel/response"
 )
 
 func (u *usecase) MaterialAll(req request.ParamModule) (*response.MaterialResponse, error) {
