+++ internal/domain/localmodel/response/location.go
